[PATCH] ext3 ordering fix
Fix the order of inodes being marked dirty in a couple of corner cases.

The only impact of this bug is that the on-disk copies of i_version
might got out of sync for a directory, or that an error inserting an
inode into a directory might leave its i_nlinks incorrect on disk for a
short interval.  Neither problem will cause trouble for ext3 during
normal operation, but the nlink problem might cause e2fsck to emit
unnecessary warnings if we crash while the incorrect version of the
inode is in the journal.