[ARM] Ensure we clear the PSR flags when calling signal handlers.
There is a problem with CPUs like the ARM926, which have Java
acceleration, which is enabled by the 'J' bit in the PSR.  Since
we used to preserve all PSR bits from user mode when calling the
signal handler, we end up calling a signal handler in Java mode.

We therefore clear all of the PSR "flags" ensuring a clean state,
and ARM or Thumb mode as appropriate when calling signal handlers.