[PATCH] make kstack_depth_to_print and some APM stuff static
Tridge has written a little utility
(http://samba.org/ftp/unpacked/junkcode/findstatic.pl) that (given a set of
object files) works out a list of candidate finctions and variables that
may be able to be made static.  Obviously some hand checking is needed
(different configs and/or architectures etc), but it seems useful.

Here is my first patch using its outout.

kstack_depth_to_print is used in many architectures, but only in one
file in each.  In some architectures it is declared and not used.  It
is not even clear why kstack_depth_to_print is not a #define.  Tridge
suggested that maybe you could change its value with a debugger ...

The APM code has a function and a struct that can be static.